+++ backend/import_wallet.go
@@ -6,8 +6,7 @@ import (
 	"fmt"
 	"strings"
 
-	"code.vegaprotocol.io/go-wallet/service"
-	"code.vegaprotocol.io/go-wallet/wallets"
+	"code.vegaprotocol.io/vegawallet/wallets"
 )
 
 type ImportWalletRequest struct {
@@ -15,6 +14,7 @@ type ImportWalletRequest struct {
 	Name       string
 	Mnemonic   string
 	Passphrase string
+	Version    uint32
 }
 
 func (r ImportWalletRequest) Check() error {
@@ -51,8 +51,7 @@ func (s *Handler) ImportWallet(data string) (ImportWalletResponse, error) {
 		return ImportWalletResponse{}, fmt.Errorf("couldn't unmarshal request: %w", err)
 	}
 
-	err = req.Check()
-	if err != nil {
+	if err = req.Check(); err != nil {
 		s.log.Errorf("Request is invalid: %v", err)
 		return ImportWalletResponse{}, fmt.Errorf("request is invalid: %w", err)
 	}
@@ -71,26 +70,7 @@ func (s *Handler) ImportWallet(data string) (ImportWalletResponse, error) {
 
 	handler := wallets.NewHandler(wStore)
 
-	svcStore, err := s.getServiceStore(config)
-	if err != nil {
-		return ImportWalletResponse{}, err
-	}
-
-	exists, err := service.ConfigExists(svcStore)
-	if err != nil {
-		s.log.Errorf("Couldn't verify service configuration existence: %v", err)
-		return ImportWalletResponse{}, fmt.Errorf("couldn't verify service configuration existence: %w", err)
-	}
-
-	if !exists {
-		err := service.GenerateConfig(svcStore, false)
-		if err != nil {
-			s.log.Errorf("Couldn't generate service configuration: %v", err)
-			return ImportWalletResponse{}, fmt.Errorf("couldn't generate service configuration: %w", err)
-		}
-	}
-
-	err = handler.ImportWallet(req.Name, req.Passphrase, req.Mnemonic)
+	err = handler.ImportWallet(req.Name, req.Passphrase, req.Mnemonic, req.Version)
 	if err != nil {
 		s.log.Errorf("Couldn't import the wallet: %v", err)
 		return ImportWalletResponse{}, err